+++ Nucleo/heartbeat/KNN.h
@@ -1,5 +1,5 @@
 // calculate the class of a set of attributes using KNN
-int performKNN(double *data, int classesNum[], double dato[], int k, int rows, int columns);
+int performKNN(float *data, int classesNum[], float dato[], int k, int rows, int columns);
 
 
 // calculate the most common type: Fashion
@@ -7,16 +7,16 @@ int mostFrequentClass(int classes[], int k);
 
 
 // extract the first N
-void extractFirstN(double data[], double firstK[], int classes[], int kClasses[], int k);
+void extractFirstN(float data[], float firstK[], int classes[], int kClasses[], int k);
 
 
 // calculate Euclidean distance between two points
-double euclideanDistance(double pt1[], double pt2[], int columns);
+float euclideanDistance(float pt1[], float pt2[], int columns);
 
 
 // calculate Euclidean distance between a point and the database
-void euclideanDistanceFromDatabase(double pt1[], double aux[], double *instances, int rows, int columns);
+void euclideanDistanceFromDatabase(float pt1[], float aux[], float *instances, int rows, int columns);
 
 
 // sort distances Ascending
-void sort(double data[], int classes[], int classesNo[], int rows);
+void sort(float data[], int classes[], int classesNo[], int rows);
